refactor: update api type files
