+++ src/app/hooks/post/useCreatePost.ts
@@ -4,7 +4,7 @@ import { useMutation } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 
 import { fetchCreatePost } from '@/app/api/post';
-import { CreatePostReq } from '@/app/types/post';
+import { CreatePostReq } from '@/app/types/api/post';
 
 export default function useCreatePost() {
   const router = useRouter();
