+++ src/app/hooks/post/usePostList.ts
@@ -4,7 +4,7 @@ import { useQuery } from '@tanstack/react-query';
 import { useEffect } from 'react';
 
 import { fetchPosts } from '@/app/api/post';
-import { PostsReq, PostsRes } from '@/app/types/post';
+import { PostsReq, PostsRes } from '@/app/types/api/post';
 
 interface Props {
   searchFilters: PostsReq;
