+++ src/app/types/api/common.ts
@@ -1,0 +1,6 @@
+interface ApiResponse {
+  ok: boolean;
+  message: string;
+}
+
+export type { ApiResponse };
