+++ src/app/types/nextAuth.d.ts
@@ -1,6 +1,6 @@
 import { DefaultSession } from 'next-auth';
 
-import { UserData } from '@/app/types/auth';
+import { UserData } from '@/app/types/api/auth';
 
 declare module 'next-auth' {
   interface Session {
